move max nside to constants
